Split as `image` submodule
